requirements.txt
discord.py
sqlalchemy
pytest